refactor(actors): naming convention
