+++ src/api/actors/effects/getActor.effect.ts
@@ -1,14 +1,14 @@
 import { Effect, HttpError, HttpStatus } from '@marblejs/core';
 import { throwError, of } from 'rxjs';
 import { mergeMap, map, catchError } from 'rxjs/operators';
-import { ActorDao } from '../model/actor.dao';
+import { ActorsDao } from '../model/actors.dao';
 import { neverNullable } from '../../../util';
-import { applyHostname } from '../model/actor.helpers';
+import { applyHostname } from '../model/actors.helpers';
 
 export const getActorEffect$: Effect = req$ =>
   req$.pipe(
     mergeMap(req => of(req.params.id).pipe(
-      mergeMap(ActorDao.findOneByImdbID),
+      mergeMap(ActorsDao.findOneByImdbID),
       mergeMap(neverNullable),
       map(applyHostname(req)),
       map(actor => ({ body: actor })),
