+++ src/api/actors/effects/getActorList.effect.spec.ts
@@ -1,17 +1,17 @@
 import * as request from 'supertest';
 import { app } from '../../../app';
 import { mockAuthorizationFor } from '../../../tests/auth.mock';
-import { mockActor } from '../../../tests/actor.mock';
+import { mockActor } from '../../../tests/actors.mock';
 import { mockUser } from '../../../tests/user.mock';
 
 describe('getActorList$', () => {
-  test('GET /api/v1/actor returns 200 and list of actors', async () => {
+  test('GET /api/v1/actors returns 200 and list of actors', async () => {
     const actors = [await mockActor(), await mockActor(), await mockActor()];
     const user = await mockUser();
     const token = await mockAuthorizationFor(user)(app);
 
     return request(app)
-      .get('/api/v1/actor')
+      .get('/api/v1/actors')
       .set('Authorization', `Bearer ${token}`)
       .expect(200)
       .then(({ body }) => {
@@ -28,19 +28,19 @@ describe('getActorList$', () => {
       });
   });
 
-  test('GET /api/v1/actor returns empty array if no actors are found', async () => {
+  test('GET /api/v1/actors returns empty array if no actors are found', async () => {
     const user = await mockUser();
     const token = await mockAuthorizationFor(user)(app);
 
     return request(app)
-      .get('/api/v1/actor')
+      .get('/api/v1/actors')
       .set('Authorization', `Bearer ${token}`)
       .expect(200, { collection: [], total: 0 });
   });
 
-  test('GET /api/v1/actor returns 401 if not authorized', async () =>
+  test('GET /api/v1/actors returns 401 if not authorized', async () =>
     request(app)
-      .get('/api/v1/actor')
+      .get('/api/v1/actors')
       .expect(401, { error: { status: 401, message: 'Unauthorized' } })
   );
 });
