+++ src/api/actors/index.apib
@@ -1,6 +1,6 @@
 # Group Actors
 
-## Actor list [/api/v1/actor{?limit,page,sortBy,sortDir}]
+## Actor list [/api/v1/actors{?limit,page,sortBy,sortDir}]
 
 ### Get all actors [GET]
 
@@ -54,7 +54,7 @@ Retrieve all actors
             }
         }
 
-## Actor [/api/v1/actor/{imdbId}]
+## Actor [/api/v1/actors/{imdbId}]
 
 ### Get single actor [GET]
 
@@ -103,4 +103,4 @@ Retrieve single actor
 + gender: male (enum, required),
     + male (string)
     + female (string)
-+ photoUrl: http://example.com/api/v1/assets/img/actor/nm0000093.jpg (string, required),
++ photoUrl: http://example.com/api/v1/assets/img/actors/nm0000093.jpg (string, required),
