+++ src/api/common/effects/getFile.effect.spec.ts
@@ -6,7 +6,7 @@ import { app } from '../../../app';
 describe('getFileEffect$', () => {
   test('GET api/v1/assets/:dir responds with 200', async () =>
     request(app)
-      .get('/api/v1/assets/img/actor/placeholder.jpg')
+      .get('/api/v1/assets/img/actors/placeholder.jpg')
       .expect(200));
 
   test('GET api/v1/assets/:dir responds with 404 if file is not found', async () =>
