+++ src/api/index.apib
@@ -24,5 +24,5 @@ Retrieve information about API availability
 
 <!-- include(auth/index.apib) -->
 <!-- include(user/index.apib) -->
-<!-- include(actor/index.apib) -->
+<!-- include(actors/index.apib) -->
 <!-- include(movie/index.apib) -->
