+++ src/tests/actors.mock.ts
@@ -1,9 +1,9 @@
 import * as faker from 'faker';
-import { ActorDao } from '../api/actor/model/actor.dao';
-import { Gender } from '../api/actor/model/actor.model';
+import { ActorsDao } from '../api/actors/model/actors.dao';
+import { Gender } from '../api/actors/model/actors.model';
 
 export const mockActor = async () =>
-  ActorDao.model.create({
+  ActorsDao.model.create({
     imdbId: faker.random.uuid(),
     name: faker.name.firstName() + ' ' + faker.name.lastName(),
     birthday: faker.date.past(),
